# Improve memory usage in track finding postamble