--- a/device/sycl/src/finding/find_tracks.hpp
+++ b/device/sycl/src/finding/find_tracks.hpp
@@ -189,6 +189,9 @@ track_candidate_container_types::buffer find_tracks(
         config.max_num_branches_per_seed * n_seeds, mr.main,
         vecmem::data::buffer_type::resizable};
     copy.setup(tips_buffer)->wait();
+    vecmem::data::vector_buffer<unsigned int> tip_length_buffer{
+        config.max_num_branches_per_seed * n_seeds, mr.main};
+    copy.setup(tip_length_buffer)->wait();
 
     std::map<unsigned int, unsigned int> step_to_link_idx_map;
     step_to_link_idx_map[0] = 0;
@@ -289,6 +292,7 @@ track_candidate_container_types::buffer find_tracks(
                      updated_liveness =
                          vecmem::get_data(updated_liveness_buffer),
                      tips = vecmem::get_data(tips_buffer),
+                     tip_lengths = vecmem::get_data(tip_length_buffer),
                      n_tracks_per_seed =
                          vecmem::get_data(n_tracks_per_seed_buffer),
                      shared_candidates_size, shared_num_candidates,
@@ -305,7 +309,7 @@ track_candidate_container_types::buffer find_tracks(
                              n_in_params, barcodes, upper_bounds, links_view,
                              prev_links_idx, curr_links_idx, step,
                              updated_params, updated_liveness, tips,
-                             n_tracks_per_seed},
+                             tip_lengths, n_tracks_per_seed},
                             {&(shared_num_candidates[0]),
                              &(shared_candidates[0]),
                              shared_candidates_size[0]});
@@ -397,7 +401,8 @@ track_candidate_container_types::buffer find_tracks(
                          param_ids = vecmem::get_data(param_ids_buffer),
                          links_view = vecmem::get_data(links_buffer),
                          prev_links_idx = step_to_link_idx_map[step], step,
-                         n_candidates, tips = vecmem::get_data(tips_buffer)](
+                         n_candidates, tips = vecmem::get_data(tips_buffer),
+                         tip_lengths = vecmem::get_data(tip_length_buffer)](
                             ::sycl::nd_item<1> item) {
                             device::propagate_to_next_surface<
                                 propagator_type,
@@ -405,7 +410,7 @@ track_candidate_container_types::buffer find_tracks(
                                 details::global_index(item), config,
                                 {det, field, in_params, param_liveness,
                                  param_ids, links_view, prev_links_idx, step,
-                                 n_candidates, tips});
+                                 n_candidates, tips, tip_lengths});
                         });
                 })
                 .wait_and_throw();
@@ -421,12 +426,16 @@ track_candidate_container_types::buffer find_tracks(
     // Get the number of tips
     auto n_tips_total = copy.get_size(tips_buffer);
 
+    std::vector<unsigned int> tips_length_host;
+
+    if (n_tips_total > 0) {
+        copy(tip_length_buffer, tips_length_host)->wait();
+        tips_length_host.resize(n_tips_total);
+    }
+
     // Create track candidate buffer
     track_candidate_container_types::buffer track_candidates_buffer{
-        {n_tips_total, mr.main},
-        {std::vector<std::size_t>(n_tips_total,
-                                  config.max_track_candidates_per_track),
-         mr.main, mr.host, vecmem::data::buffer_type::resizable}};
+        {n_tips_total, mr.main}, {tips_length_host, mr.main, mr.host}};
     copy.setup(track_candidates_buffer.headers)->wait();
     copy.setup(track_candidates_buffer.items)->wait();
     track_candidate_container_types::view track_candidates =