The code which turns the tips of our track finding into actual tracks uses an excessive amount of memory, as it massively overallocates. Indeed, it allocates memory as though all tips have the maximum number of track states, which is unrealistic. This commit makes it so that the number of valid track states is counted instead, making the allocation more precise. In my measurements, this more than halves the memory usage of traccc on $\langle\mu\rangle = 200$ ttbar events in the ODD.

### Before:

![6fee5ca3](https://github.com/user-attachments/assets/45915974-cf70-4da2-95e7-43afc2e5483c)

```
Reconstructed track parameters: 884796
Time totals:
                  File reading  5202 ms
              Event processing  4388 ms
Throughput:
              Event processing  438.829 ms/event, 2.27879 events/s
```

### After:

![a91e94d0](https://github.com/user-attachments/assets/ee323394-a0d9-4f6b-a7b7-85918b12b364)


```
Reconstructed track parameters: 884784
Time totals:
                  File reading  5128 ms
              Event processing  4352 ms
Throughput:
              Event processing  435.297 ms/event, 2.29728 events/s
```

Note: The SYCL changes are best-effort but I haven't tested them.