@stephenswat 2025-03-13
I was able to one-up myself, removing the counting kernel entirely, reducing the memory allocations, keeping the memory usage low, and improving performance:

```
Reconstructed track parameters: 884809
Time totals:
                  File reading  5144 ms
              Event processing  4230 ms
Throughput:
              Event processing  423.022 ms/event, 2.36394 events/s
```