@stephenswat 2025-03-17
So the responsibility of the kernels in the old version of the algorithm was as follows:

* `propagate_to_next_surface` propagates the tracks and creates _tips_, i.e. points where a track might have ended. Each tracks corresponds 1:1 with a surface link, which describes the head of a reversed linked list of candidates.
* `build_tracks` follows the linked list of candidates starting at each tip, computes some properties about that track, and places _all_ tracks (regardless of whether they make sense or not) in a very large jagged array
* `prune_tracks` looks at the reified tracks in the big jagged vector and copies the good ones (based on the properties determined by the `build_tracks` kernel) into the final jagged vector.

Now, the core insight is that the checks that `build_tracks` performs are extremely rudimentary. It's literally just that the number of non-hole candidates is between a lower and an upper bound. All other cuts are already performed by the `propagate_to_next_surface`. Thus, the core of this PR is to move the checks that `build_tracks` does to `propagate_to_next_surface`, which has all the necessary information already. The core idea is that we create only tips for valid tracks, rather than tips for _all_ possible tracks. This, of course, also makes it trivial to count the number of valid tracks and how long those tracks will be. So in the new model, the tasks are distributed as follows:

* `propagate_to_next_surface` propagates the tracks and creates tips _only for valid tracks_, where the validity is defined by some very simple set of metrics. The definition of tips remain the same, but we create far fewer of them, and we record the length of each track.
* `build_tracks` no longer does any filtering, it just follows the linked list and places track states into a much smaller jagged vector.

Much simpler, and it has far fewer kernels. Note that in this new model `build_tracks` basically does what `prune_tracks` did in the previous model.