@beomki-yeo 2025-03-17
> The core idea is that we create only tips for valid tracks, rather than tips for all possible tracks. 

Yeah I think this is a right way to improve :thinking: