@stephenswat 2025-04-02
# Performance summary
Here is a summary of the performance effects of this PR:
## Graphical
![](https://traccc-bench.web.cern.ch/908_921c876887be26f87ef9261c45702c77a52ffae3.png)
## Tabular
| Kernel | 348de1a2 | 921c8768 | Delta |
| --- | --: | --: | --: |
| `fit` | 277.88 ms | 256.77 ms | -7.6% |
| `propagate_to_next_surface` | 118.17 ms | 117.00 ms | -1.0% |
| `find_tracks` | 26.65 ms | 23.00 ms | -13.7% |
| `count_triplets` | 14.16 ms | 14.16 ms | -0.0% |
| `find_triplets` | 5.98 ms | 5.98 ms | 0.0% |
| `find_doublets` | 911.31 μs | 893.99 μs | -1.9% |
| `Thrust::sort` | 757.86 μs | 758.34 μs | 0.1% |
| `ccl_kernel` | 682.50 μs | 682.54 μs | 0.0% |
| `count_doublets` | 627.27 μs | 628.01 μs | 0.1% |
| `build_tracks` | 1.10 ms | 445.35 μs | -59.6% |
| `prune_tracks` | 1.23 ms | &mdash; | &mdash; |
| `select_seeds` | 362.74 μs | 358.26 μs | -1.2% |
| `apply_interaction` | 132.45 μs | 131.94 μs | -0.4% |
| `update_triplet_weights` | 96.16 μs | 96.23 μs | 0.1% |
| `fill_sort_keys` | 68.25 μs | 67.72 μs | -0.8% |
| `estimate_track_params` | 34.31 μs | 34.31 μs | 0.0% |
| `populate_grid` | 30.40 μs | 30.45 μs | 0.2% |
| `count_grid_capacities` | 29.21 μs | 29.18 μs | -0.1% |
| `form_spacepoints` | 14.24 μs | 14.20 μs | -0.3% |
| `reduce_triplet_counts` | 6.71 μs | 6.71 μs | 0.0% |
| `static_kernel` | 1.78 μs | 1.78 μs | 0.1% |
| `make_barcode_sequence` | 992.11 ns | 1.00 μs | 1.3% |
| `fill_prefix_sum` | 165.57 ns | 165.51 ns | -0.0% |
| **Total** | **448.93 ms** | **421.08 ms** | **-6.2%** |
> [!NOTE]
> This is an automated message produced on the explicit request of a human being.